collectors: only export aggregated DescribeGraph metrics
In this commit, we modify the way we export the graph metrics to no
longer export each individual piece of information. It has been
observed, that with our current scrape interval, the amount of data can
quickly exceed what can be held in memory on a box with 4GB of memory.
As a result, in order to reduce the amount of memory used, we'll now
just export the min/max/avg/median of fields like the time lock delta of
channel. This change makes the queries more accessible, and should also
reduce the memory usage of `lndmon` as a whole.