dashboards: more accurate CPU usage for the perf dashboard
In this PR, we attempt to make the current CPU usage graph for `lnd`
more accurate. AFAICT, the `*10` that we had was completely superfluous,
and omitting that seems to produce a more accurate CPU usage reading. In
the future, we may want to use something like [`Node
Exporter`](https://prometheus.io/docs/guides/node-exporter/) as we can
obtain richer information such as the idle CPU time, or the amount of
disk space left on the instance.