dashboards: switch to irate for peer bytes sent/recv
In this commit, we switch to using an `irate` over the past minute for
the peer graphs. The cumulative graph wasn't very helpful, since it can
be skewed when peers connect/reconnect. This new `irate` format is
better as you can at a glance, spot anomalies (spikes in sent/recv)
easily.

We also fix some graphs that were graph satoshis instead of bytes (kek),
and use a longer interval for the peer flappiness graph to make it
smoother.