config: allow specifying of macaroon directory and name
Previous changes set lndmon to only use the macaroon path that is set
when we use docker-compose and the mounted macaroons volume. This commit
updates lndmon to use the macaroon dir path and adds a macaroon name (so
that macaroons other than the readonly macaroon could be used). This
allows standalone lndmon containers that do not use our docker-compose
setup to specify where their macaroons are.