collectors: add peer label
Adding the peer label makes it easier to monitor or alert on conditions
that are specific to a node with a known pubkey. This is particularly
useful if one is running multiple nodes that are connected with internal
channels. It is often the case that different conditions apply to these
connections.